refactor: extract shared search shaping helpers into their own modules (#65, #48)
Mechanical move ahead of the search/grep verb split. Shaping, paging,
rendering, and FTS retrieval prep move from commands/search.rs to
commands/search_common.rs; the meeting-filter predicates move to
query/filter.rs so retrieval code can reach them without depending on
the command layer. No behavior change.
